requirements.txt
click==7.1.1
cycler==0.10.0
decorator==4.4.2
Flask==1.1.2
gunicorn==20.0.4
imageio==2.8.0
imutils==0.5.3
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
networkx==2.4
numpy==1.18.3
opencv-python==4.2.0.34
Pillow==7.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.16.2
scipy==1.4.1
six==1.14.0
Werkzeug==1.0.1